:recycle: api: move CertificationCandidateForbiddenDeletionError to src
