+++ api/src/certification/session/domain/errors.js
@@ -37,10 +37,17 @@ class SessionWithMissingAbortReasonError extends DomainError {
   }
 }
 
+class CertificationCandidateForbiddenDeletionError extends DomainError {
+  constructor(message = 'Il est interdit de supprimer un candidat de certification déjà lié à un utilisateur.') {
+    super(message);
+  }
+}
+
 export {
   SessionStartedDeletionError,
   SessionWithMissingAbortReasonError,
   SessionWithoutStartedCertificationError,
   SessionWithAbortReasonOnCompletedCertificationCourseError,
   SessionAlreadyFinalizedError,
+  CertificationCandidateForbiddenDeletionError,
 };
