+++ api/src/shared/application/error-manager.js
@@ -7,8 +7,11 @@ import { extractLocaleFromRequest } from '../../../lib/infrastructure/utils/requ
 import _ from 'lodash';
 import * as translations from '../../../translations/index.js';
 import { AdminMemberError } from '../../authorization/domain/errors.js';
-import { SessionStartedDeletionError } from '../../certification/session/domain/errors.js';
 import { domainErrorMapper } from './domain-error-mapper.js';
+import {
+  SessionStartedDeletionError,
+  CertificationCandidateForbiddenDeletionError,
+} from '../../certification/session/domain/errors.js';
 
 const { Error: JSONAPIError } = jsonapiSerializer;
 const NOT_VALID_RELATIONSHIPS = ['externalId', 'participantExternalId'];
@@ -108,6 +111,9 @@ function _mapToHttpError(error) {
   }
   if (error instanceof SessionStartedDeletionError) {
     return new HttpErrors.ConflictError(error.message);
+  }
+  if (error instanceof CertificationCandidateForbiddenDeletionError) {
+    return new HttpErrors.ForbiddenError(error.message);
   }
 
   return new HttpErrors.BaseHttpError(error.message);
